Backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
beautifulsoup4
faiss-cpu
sentence-transformers
httpx
pydantic
